// app/src/api-client/api.ts
export const TaskState = {
    Active: 'Active',
    Finished: 'Finished'
} as const;

export type TaskState = typeof TaskState[keyof typeof TaskState];


export interface TodoTaskDto {
    'name': string;
    'description': string;
    'estimatedTime': number;
}
export interface TodoTaskDtoRecord {
    'id': number;
    'name': string;
    'description': string;
    'estimatedTime': number;
    'state': TaskState;
    'createdAt': string;
    'updatedAt': string;
    'order': number;
    'finishedAt'?: string | null;
}